+++ ccip-sdk/src/requests.ts
@@ -198,7 +198,7 @@ const BLOCK_LOG_WINDOW_SIZE = 5000
  * @param opts - Optional log filtering parameters.
  * @returns Array of messages in the batch.
  */
-export async function fetchAllMessagesInBatch<
+export async function getAllMessagesInBatch<
   C extends Chain,
   R extends PickDeep<
     CCIPRequest,