+++ ccip-sdk/src/solana/index.ts
@@ -115,7 +115,7 @@ import {
   simulateAndSendTxs,
   simulationProvider,
 } from './utils.ts'
-import { fetchAllMessagesInBatch, getMessageById, getMessagesInTx } from '../requests.ts'
+import { getAllMessagesInBatch, getMessageById, getMessagesInTx } from '../requests.ts'
 import { patchBorsh } from './patchBorsh.ts'
 
 const routerCoder = new BorshCoder(CCIP_ROUTER_IDL)
@@ -436,8 +436,8 @@ export class SolanaChain extends Chain<typeof ChainFamily.Solana> {
     return getMessageById(this, messageId, { address: onRamp, ...opts })
   }
 
-  /** {@inheritDoc Chain.fetchAllMessagesInBatch} */
-  async fetchAllMessagesInBatch<
+  /** {@inheritDoc Chain.getAllMessagesInBatch} */
+  async getAllMessagesInBatch<
     R extends PickDeep<
       CCIPRequest,
       'lane' | `log.${'topics' | 'address' | 'blockNumber'}` | 'message.sequenceNumber'
@@ -451,14 +451,14 @@ export class SolanaChain extends Chain<typeof ChainFamily.Solana> {
       [Buffer.from('dest_chain_state'), toLeArray(request.lane.destChainSelector, 8)],
       new PublicKey(request.log.address),
     )
-    // fetchAllMessagesInBatch pass opts back to getLogs; use it to narrow getLogs filter only to
+    // getAllMessagesInBatch pass opts back to getLogs; use it to narrow getLogs filter only to
     // txs touching destChainStatePda
     const opts_: Parameters<SolanaChain['getLogs']>[0] = {
       ...opts,
       programs: [request.log.address],
       address: destChainStatePda.toBase58(),
     }
-    return fetchAllMessagesInBatch(this, request, commit, opts_)
+    return getAllMessagesInBatch(this, request, commit, opts_)
   }
 
   /** {@inheritDoc Chain.typeAndVersion} */